requirements.txt
black==25.1.0
Django==5.1.6
gunicorn==23.0.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
virtualenv==20.29.2
python-decouple==3.8
djangorestframework==3.15.2
drf-yasg==1.21.8
